Replace old test file with updated TypeScript test for `<TableToolbar />` custom icons
